server/requirements.txt
fastapi==0.88.0
uvicorn==0.20.0
tensorflow
numpy==1.24.3
pillow==9.5.0
python-multipart